Migrate index.ts to Typescript
